+++ src/components/Sertifikat/SertifikatList.jsx
@@ -3,6 +3,8 @@ import axios from "axios";
 import { Link } from "react-router-dom";
 import { useSelector } from "react-redux";
 import Modal from "../Modal";
+import Swal from "sweetalert2";
+import { IoTrashOutline, IoCreateOutline } from "react-icons/io5";
 
 const SertifikatList = () => {
   const [sertifikat, setSertifikat] = useState([]);
@@ -21,8 +23,26 @@ const SertifikatList = () => {
   };
 
   const deleteSertifikat = async (sertifikatId) => {
-    await axios.delete(`http://localhost:5000/sertifikat/${sertifikatId}`);
-    getSertifikat();
+    Swal.fire({
+      title: "Apakah Anda Yakin Akan Menghapus Data Ini?",
+      text: "Data arsip sertifikat yang terhapus tidak akan bisa kembali!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Ya, Hapus",
+      cancelButtonText: "Batal",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await axios.delete(`http://localhost:5000/sertifikat/${sertifikatId}`);
+        getSertifikat();
+        Swal.fire(
+          "Terhapus!",
+          "Data Arsip Sertifikat Telah Dihapus.",
+          "success"
+        );
+      }
+    });
   };
 
   const openModal = (arsipSertifikat) => {
@@ -40,7 +60,7 @@ const SertifikatList = () => {
 
   return (
     <div>
-      <h1 className="title">Kelola Arsip Sertifikat</h1>
+      <h1 className="title mt-2">Kelola Arsip Sertifikat</h1>
       <h2 className="subtitle">Daftar Data Sertifikat</h2>
       <div className="container mb-2">
         <div className="control">
@@ -90,7 +110,13 @@ const SertifikatList = () => {
                   <td>
                     <button
                       className="button is-small is-primary"
-                      onClick={() => openModal(sertifikat.arsip_sertifikat)}
+                      onClick={() =>
+                        openModal(
+                          "https://" +
+                            sertifikat.arsip_sertifikat +
+                            ".ipfs.w3s.link"
+                        )
+                      }
                     >
                       Arsip Sertifikat
                     </button>
@@ -108,13 +134,13 @@ const SertifikatList = () => {
                         to={`/sertifikat/edit/${sertifikat.uuid}`}
                         className="button is-small is-info"
                       >
-                        Ubah
+                        <IoCreateOutline />
                       </Link>
                       <button
                         onClick={() => deleteSertifikat(sertifikat.uuid)}
                         className="button is-small is-danger"
                       >
-                        Hapus
+                        <IoTrashOutline />
                       </button>
                     </td>
                   )}
